split up wrapper binding classes; fix circular references in finalizers; add unit tests
